pyproject.toml
[tool.poetry]
name = "tfds-project"
version = "0.1.0"
description = ""
authors = ["dorabijvoet <dora.bijvoet@hotmail.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
langchain-community = "^0.3.0"
sentence-transformers = "^3.1.1"
rerankers = "^0.5.3"
langchain-chroma = "^0.1.4"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
notebook = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
